api: split DuckDBType.ts into one file per type class
Move each type class (and its companion singleton const / factory) into its
own file under api/src/types/, matching the existing api/src/values/ layout.
The abstract base lives in BaseDuckDBType.ts and the DuckDBType union in
types/DuckDBType.ts. api/src/DuckDBType.ts becomes a re-export of ./types so
all existing import paths keep working unchanged.

Pure structural change: no behavior or public API changes.
